# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Participium API - Citizen Registration & Municipality Administration
  description: |
    API for citizen registration, authentication (PT01), municipality user management (PT02) and municipality users role assignment (PT03).
    Uses session-based authentication with HTTP-only cookies.

    ## Municipality User Management
    System administrators can create, update and delete internal municipality user accounts.
    The system must have at least one administrator account pre-configured in the database.

    Available municipality roles:
    - ADMINISTRATOR: System administrator with full privileges
    - PUBLIC_RELATIONS: Public relations officer for citizen communication
    - CULTURE_EVENTS_TOURISM_SPORTS: Culture, events, tourism and sports
    - LOCAL_PUBLIC_SERVICES: Local public services
    - EDUCATION_SERVICES: Education services
    - PUBLIC_RESIDENTIAL_HOUSING: Public/residential housing services
    - INFORMATION_SYSTEMS: Information systems/IT
    - MUNICIPAL_BUILDING_MAINTENANCE: Municipal building maintenance
    - PRIVATE_BUILDINGS: Private buildings oversight
    - INFRASTRUCTURES: Road, bridges and general infrastructures
    - GREENSPACES_AND_ANIMAL_PROTECTION: Parks, green areas and animal protection
    - WASTE_MANAGEMENT: Waste collection and management
    - ROAD_MAINTENANCE: Road maintenance
    - CIVIL_PROTECTION: Civil protection and emergency coordination
    
    External roles:
...
  version: 1.1.0
  contact:
...
servers:
...
paths:
...
  /admin/external-companies:
    post:
      tags:
        - Administration
      summary: Create a new external company
      description: |
        Create a new external company that can potentially be assigned report maintenance tasks.
        Only administrators can perform this operation.
      operationId: createExternalCompany
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - categories
                - platformAccess
              properties:
                name:
                  type: string
                  description: Company name
                  example: "Green Tech Solutions"
                categories:
                  type: array
                  items:
                    $ref: "#/components/schemas/ReportCategory"
                  minItems: 1
                  maxItems: 2
                  description: Report categories this company can handle (max 2)
                  example: ["GREENSPACES", "ROAD_MAINTENANCE"]
                platformAccess:
                  type: boolean
                  description: Whether this company can have maintainer accounts
                  example: true
      responses:
        "201":
          description: External company created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: "Green Tech Solutions"
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportCategory"
                    example: ["GREENSPACES", "ROAD_MAINTENANCE"]
                  platformAccess:
                    type: boolean
                    example: true
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidName:
                  value:
                    code: 400
                    error: "BadRequest" 
                    message: "Company name is required"
                invalidCategories:
                  value:
                    code: 400
                    error: "BadRequest"
                    message: "Maximum 2 categories allowed"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                error: "Unauthorized"
                message: "Authentication required"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 403
                error: "Forbidden"
                message: "Administrator privileges required"
    
    get:
      tags:
        - Administration
      summary: List all external companies
      description: |
        Returns all external companies with their associated maintainer accounts.
        Only administrators can perform this operation.
      operationId: listExternalCompanies
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Array of external companies with maintainers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    categories:
                      type: array
                      items:
                        $ref: "#/components/schemas/ReportCategory"
                    platformAccess:
                      type: boolean
                    maintainers:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          firstName:
                            type: string
                          lastName:
                            type: string
                          email:
                            type: string
                          role:
                            type: string
              example:
                - id: 1
                  name: "Green Tech Solutions"
                  categories: ["GREENSPACES", "ROAD_MAINTENANCE"]
                  platformAccess: true
                  maintainers:
                    - id: 101
                      firstName: "Mario"
                      lastName: "Rossi"
                      email: "mario.rossi@greentech.com"
                      role: "EXTERNAL_MAINTAINER"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                error: "Unauthorized"
                message: "Authentication required"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 403
                error: "Forbidden"
                message: "Administrator privileges required"
  /admin/external-companies/{id}:
    delete:
      tags:
        - Administration
      summary: Delete an external company
      description: |
        Delete an external company by ID. The company must not have any associated maintainers.
        Only administrators can perform this operation.
      operationId: deleteExternalCompany
      security:
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: External company ID to delete
          example: 1
      responses:
        "204":
          description: External company deleted successfully
        "400":
          description: Invalid input or business rule violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidId:
                  value:
                    code: 400
                    error: "BadRequest"
                    message: "Invalid company ID parameter"
                hasMaintainers:
                  value:
                    code: 400
                    error: "BadRequest"
                    message: "Cannot delete company with existing maintainers. Remove all maintainers first."
                deleteFailed:
                  value:
                    code: 400
                    error: "BadRequest"
                    message: "Failed to delete external company"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                error: "Unauthorized"
                message: "Authentication required"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 403
                error: "Forbidden"
                message: "Administrator privileges required"
        "404":
          description: External company not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 404
                error: "NotFound"
                message: "External company not found"
  /admin/external-companies/platform-access:
    get:
      tags:
        - Administration
      summary: Get external companies with platform access
      description: |
        Returns external companies that have platform access and can have maintainer accounts created.
        Used for dropdown/selection UI when creating external maintainers.
      operationId: getExternalCompaniesWithAccess
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Companies with platform access
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    categories:
                      type: array
                      items:
                        $ref: "#/components/schemas/ReportCategory"
                    platformAccess:
                      type: boolean
              example:
                - id: 1
                  name: "Green Tech Solutions"
                  categories: ["GREENSPACES", "ROAD_MAINTENANCE"]
                  platformAccess: true
        "401":
          description: Not authenticated
        "403":
          description: Insufficient permissions
  /admin/external-maintainers:
    post:
      tags:
        - Administration
      summary: Create an external maintainer account
      description: |
        Create a new external maintainer account linked to a company with platform access.
        Only administrators can perform this operation.
      operationId: createExternalMaintainer
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
                - email
                - password
                - externalCompanyId
              properties:
                firstName:
                  type: string
                  example: "Mario"
                lastName:
                  type: string
                  example: "Rossi"
                email:
                  type: string
                  format: email
                  example: "mario.rossi@greentech.com"
                password:
                  type: string
                  minLength: 8
                  example: "SecurePass123!"
                externalCompanyId:
                  type: string
                  description: ID of external company (as string for form compatibility)
                  example: "1"
      responses:
        "201":
          description: External maintainer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  company:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      categories:
                        type: array
                        items:
                          $ref: "#/components/schemas/ReportCategory"
                      platformAccess:
                        type: boolean
              example:
                id: 101
                firstName: "Mario"
                lastName: "Rossi"
                email: "mario.rossi@greentech.com"
                role: "EXTERNAL_MAINTAINER"
                company:
                  id: 1
                  name: "Green Tech Solutions"
                  categories: ["GREENSPACES", "ROAD_MAINTENANCE"]
                  platformAccess: true
        "400":
          description: Invalid input or business rule violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                emailExists:
                  value:
                    code: 400
                    error: "BadRequest"
                    message: "Email already exists"
                noPlatformAccess:
                  value:
                    code: 400
                    error: "BadRequest"
                    message: "Cannot create maintainer for company without platform access"
                weakPassword:
                  value:
                    code: 400
                    error: "BadRequest"
                    message: "password is required and must be at least 8 characters"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient permissions
        "404":
          description: External company not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 404
                error: "NotFound"
                message: "External company not found"
...
  /reports/{reportId}/assignable-externals:
    get:
      tags:
        - Reports
      summary: List external companies and maintainers available for this report's category
      description: |
        Return the list of external companies and their maintainer users that can handle the given report.
        The backend filters available external companies according to the report's category (e.g., Enel X for PUBLIC_LIGHTING).
        
        Returns two types of assignees:
        - **Users**: For companies with platform access (has_platform_access=true), returns individual maintainer users
        - **Companies**: For companies without platform access (has_platform_access=false), returns only company information
        
        This endpoint is intended for use by technical office staff members when they need to assign a report to an external entity.
      operationId: getAssignableExternals
      security:
        - SessionAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          description: ID of the report
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Array of external companies that can handle this report category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExternalCompanyResponse"
              example:
                - id: 10
                  name: Enel X
                  categories:
                    - PUBLIC_LIGHTING
                  platformAccess: true
                  users:
                    - id: 501
                      firstName: Marco
                      lastName: Bianchi
                      email: marco.bianchi@enelx.com
                      role: EXTERNAL_MAINTAINER
                    - id: 502
                      firstName: Giulia
                      lastName: Ferrari
                      email: giulia.ferrari@enelx.com
                      role: EXTERNAL_MAINTAINER
                - id: 11
                  name: AMIAT Torino
                  categories:
                    - WASTE
                    - ROADS_URBAN_FURNISHINGS
                  platformAccess: false
                  users: null
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 401
                error: "Unauthorized"
                message: "Authentication required"
        "403":
          description: Insufficient permissions (must be assigned technical officer)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notTechnicalStaff:
                  summary: User does not have technical staff role
                  value:
                    code: 403
                    error: "Forbidden"
                    message: "Technical office staff privileges required"
                notAssignedOfficer:
                  summary: User is not the assigned technical officer for this report
                  value:
                    code: 403
                    error: "Forbidden"
                    message: "Only the assigned technical officer can view assignable externals"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 404
                error: "NotFound"
                message: "Report not found"
        "500":
          description: Unable to retrieve external maintainers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: 500
                error: "InternalServerError"
                message: "Failed to retrieve assignable external maintainers"
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: |
        Session-based authentication using HTTP-only cookies with Passport.js.
        The session cookie is automatically sent by the browser with each request.
        Frontend must use 'credentials: include' in fetch/axios requests.
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
      properties:
        code:
          type: integer
          description: HTTP status code of the error
          example: 400
        error:
          type: string
          description: Error type (BadRequest, Unauthorized, Conflict, InternalServerError)
          example: "BadRequest"
        message:
          type: string
          description: Human-readable error message
          example: "Missing required fields: email, password"
...
    ExternalCompanyResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the external company
          example: 10
        name:
          type: string
          description: Name of the external company
          example: Enel X
        categories:
          type: array
          items:
            $ref: "#/components/schemas/ReportCategory"
          description: List of report categories this company can handle (max 2)
          minItems: 1
          maxItems: 2
          example:
            - PUBLIC_LIGHTING
        platformAccess:
          type: boolean
          description: Whether the company has access to the Participium platform
          example: true
...
    ReportCategory:
      type: string
      description: Category of the reported problem
      enum:
        - WATER_SUPPLY_DRINKING_WATER
        - ARCHITECTURAL_BARRIERS
        - SEWER_SYSTEM
        - PUBLIC_LIGHTING
        - WASTE
        - ROAD_SIGNS_TRAFFIC_LIGHTS
        - ROADS_URBAN_FURNISHINGS
        - PUBLIC_GREEN_AREAS_PLAYGROUNDS
        - OTHER
      example: ROADS_URBAN_FURNISHINGS
...